backend/requirements.txt
# HandOff.AI Backend — Python Dependencies

# Web framework & async server
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==13.1
httpx==0.27.2

# Fetch.ai uAgents multi-agent framework (latest stable)
uagents==0.23.7
uagents-core>=0.4.3

# Firebase Admin SDK (Firestore + Realtime Database)
firebase-admin==6.5.0

# Google Gemini Vision API
google-generativeai==0.8.3

# Data validation & serialization
pydantic==2.9.2
pydantic-settings==2.6.1

# Environment & configuration
python-dotenv==1.0.1

# Async utilities
anyio==4.6.2
aiofiles==24.1.0

# JWT auth (Firebase token verification)
PyJWT==2.9.0

# Image processing (screenshot encoding)
Pillow==11.0.0

# Retry / resilience
tenacity==8.5.0

# Numeric (cosine similarity for semantic SOP search)
numpy>=1.26.4

# Logging
structlog==24.4.0
